# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/master-data/countries/{isoCode}:
    get:
      tags:
      - country-controller
      operationId: getCountry
      parameters:
      - name: isoCode
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CountryResponse'
    put:
      tags:
      - country-controller
      operationId: updateCountry
      parameters:
      - name: isoCode
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCountryRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CountryResponse'
  /api/master-data/countries/{isoCode}/deactivate:
    put:
      tags:
      - country-controller
      operationId: deactivateCountry
      parameters:
      - name: isoCode
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CountryResponse'
  /api/master-data/channels/{code}:
    put:
      tags:
      - channel-controller
      operationId: updateChannel
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChannelRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ChannelResponse'
  /api/master-data/channels/{code}/deactivate:
    put:
      tags:
      - channel-controller
      operationId: deactivateChannel
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ChannelResponse'
  /api/master-data/branches/{code}:
    get:
      tags:
      - branch-controller
      operationId: getBranch
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BranchResponse'
    put:
      tags:
      - branch-controller
      operationId: updateBranch
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBranchRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BranchResponse'
  /api/master-data/branches/{code}/deactivate:
    put:
      tags:
      - branch-controller
      operationId: deactivateBranch
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BranchResponse'
...
  /api/customers/{id}:
    get:
      tags:
      - customer-controller
      summary: Get customer by ID
      description: Retrieve customer details by internal database ID
      operationId: getCustomerById
      parameters:
      - name: id
        in: path
        description: Customer database ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponse'
    put:
      tags:
      - customer-controller
      summary: Update customer
      description: Update customer details. Supports partial updates.
      operationId: updateCustomer
      parameters:
      - name: id
        in: path
        description: Customer database ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponse'
...
  /api/master-data/countries:
    get:
      tags:
      - country-controller
      operationId: getAllCountries
      parameters:
      - name: activeOnly
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'
    post:
      tags:
      - country-controller
      operationId: createCountry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCountryRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CountryResponse'
  /api/master-data/channels:
    get:
      tags:
      - channel-controller
      operationId: getActiveChannels
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChannelResponse'
    post:
      tags:
      - channel-controller
      operationId: createChannel
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ChannelResponse'
  /api/master-data/branches:
    get:
      tags:
      - branch-controller
      operationId: getActiveBranches
      parameters:
      - name: countryCode
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BranchResponse'
    post:
      tags:
      - branch-controller
      operationId: createBranch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBranchRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BranchResponse'
...
  /api/customers/sme:
    post:
      tags:
      - customer-controller
      summary: Create a new SME customer
      description: Creates a new small/medium enterprise customer with business type
        and turnover
      operationId: createSME
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSMECustomerRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponse'
  /api/customers/individual:
    post:
      tags:
      - customer-controller
      summary: Create a new individual customer
      description: Creates a new individual customer with personal details
      operationId: createIndividual
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIndividualCustomerRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponse'
  /api/customers/corporate:
    post:
      tags:
      - customer-controller
      summary: Create a new corporate customer
      description: "Creates a new corporate customer with registration details, industry,\
        \ and annual revenue"
      operationId: createCorporate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCorporateCustomerRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponse'
...
  /api/limits/assignments/customer/{id}:
    get:
      tags:
      - limit-assignment-controller
      operationId: getCustomerLimits
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerLimitResponse'
...
  /api/customers:
    get:
      tags:
      - customer-controller
      summary: List all customers
      description: Returns paginated list of all customers (no filtering)
      operationId: listCustomers
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCustomerResponse'
  /api/customers/search:
    get:
      tags:
      - customer-controller
      summary: Search customers
      description: "Search customers by criteria (name, type, status, etc.)"
      operationId: searchCustomers
      parameters:
      - name: criteria
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/CustomerSearchCriteria'
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCustomerResponse'
...
  /api/charges/assignments/customer/{customerId}:
    get:
      tags:
      - charge-assignment-controller
      operationId: getCustomerOverrides
      parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerChargeOverrideResponse'
  /api/authorizations/customer/{customerId}:
    get:
      tags:
      - authorization-controller
      operationId: getAuthorizationsByCustomer
      parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuthorizationResponse'
...
components:
  schemas:
...
    UpdateCountryRequest:
      type: object
      properties:
        name:
          type: string
        region:
          type: string
    CountryResponse:
      type: object
      properties:
        isoCode:
          type: string
        name:
          type: string
        region:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateChannelRequest:
      type: object
      properties:
        name:
          type: string
    ChannelResponse:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateBranchRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
    BranchResponse:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        countryCode:
          type: string
        address:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateCustomerRequest:
      required:
      - name
      type: object
      properties:
        name:
          type: string
        taxId:
          type: string
        registrationNumber:
          type: string
        industry:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        annualRevenueAmount:
          type: integer
          format: int64
        annualRevenueCurrency:
          type: string
        employeeCount:
          type: integer
          format: int32
        website:
          type: string
        businessType:
          type: string
        annualTurnoverAmount:
          type: integer
          format: int64
        annualTurnoverCurrency:
          type: string
        yearsInOperation:
          type: integer
          format: int32
        dateOfBirth:
          type: string
          format: date
        placeOfBirth:
          type: string
        nationality:
          type: string
        employerId:
          type: integer
          format: int64
        employmentStatus:
          type: string
    AddressDto:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerNumber:
          type: string
        type:
          type: string
          enum:
          - CORPORATE
          - SME
          - INDIVIDUAL
        status:
          type: string
          enum:
          - PENDING
          - ACTIVE
          - SUSPENDED
          - CLOSED
        name:
          type: string
        taxId:
          type: string
        registrationNumber:
          type: string
        industry:
          type: string
        annualRevenueAmount:
          type: number
        annualRevenueCurrency:
          type: string
        employeeCount:
          type: integer
          format: int32
        website:
          type: string
        businessType:
          type: string
        annualTurnoverAmount:
          type: number
        annualTurnoverCurrency:
          type: string
        yearsInOperation:
          type: integer
          format: int32
        dateOfBirth:
          type: string
          format: date
        placeOfBirth:
          type: string
        nationality:
          type: string
        employerId:
          type: integer
          format: int64
        employmentStatus:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - TERMINATED
          - ON_LEAVE
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/AddressDto'
        phones:
          type: array
          items:
            $ref: '#/components/schemas/PhoneDto'
        emails:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
    PhoneDto:
      type: object
      properties:
        countryCode:
          type: string
        number:
          type: string
        type:
          type: string
          enum:
          - MOBILE
          - LANDLINE
          - WORK
          - FAX
...
    AuthorizationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        customerName:
          type: string
        authorizedPersonId:
          type: integer
          format: int64
        authorizedPersonName:
          type: string
        relationshipType:
          type: string
          enum:
          - OWNER
          - DIRECTOR
          - AUTHORIZED_SIGNATORY
          - BENEFICIAL_OWNER
        isPrimary:
          type: boolean
        effectiveDate:
          type: string
          format: date
        expirationDate:
          type: string
          format: date
        status:
          type: string
          enum:
          - ACTIVE
          - EXPIRED
          - REVOKED
        documentReferences:
          type: array
          items:
            type: string
...
    CreateCountryRequest:
      required:
      - isoCode
      - name
      type: object
      properties:
        isoCode:
          type: string
        name:
          type: string
        region:
          type: string
    CreateChannelRequest:
      required:
      - code
      - name
      type: object
      properties:
        code:
          type: string
        name:
          type: string
    CreateBranchRequest:
      required:
      - code
      - countryCode
      - name
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        countryCode:
          type: string
        address:
          type: string
...
    CustomerLimitResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        limitDefinitionId:
          type: integer
          format: int64
        limitName:
          type: string
        customerId:
          type: integer
          format: int64
        overrideAmount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateSMECustomerRequest:
      required:
      - name
      - taxId
      type: object
      properties:
        name:
          type: string
        taxId:
          type: string
        registrationNumber:
          type: string
        industry:
          type: string
        businessType:
          type: string
        annualTurnoverAmount:
          type: string
        annualTurnoverCurrency:
          type: string
        yearsInOperation:
          type: integer
          format: int32
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/AddressDto'
        phones:
          type: array
          items:
            $ref: '#/components/schemas/PhoneDto'
        emails:
          type: array
          items:
            type: string
    CreateIndividualCustomerRequest:
      required:
      - name
      - taxId
      type: object
      properties:
        name:
          type: string
        taxId:
          type: string
        dateOfBirth:
          type: string
          format: date
        placeOfBirth:
          type: string
        nationality:
          type: string
        employerId:
          type: integer
          format: int64
        employmentStatus:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - TERMINATED
          - ON_LEAVE
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/AddressDto'
        phones:
          type: array
          items:
            $ref: '#/components/schemas/PhoneDto'
        emails:
          type: array
          items:
            type: string
    CreateCorporateCustomerRequest:
      required:
      - name
      - taxId
      type: object
      properties:
        name:
          type: string
        taxId:
          type: string
        registrationNumber:
          type: string
        industry:
          type: string
        annualRevenueAmount:
          type: string
        annualRevenueCurrency:
          type: string
        employeeCount:
          type: integer
          format: int32
        website:
          type: string
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/AddressDto'
        phones:
          type: array
          items:
            $ref: '#/components/schemas/PhoneDto'
        emails:
          type: array
          items:
            type: string
...
    CustomerChargeOverrideResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        chargeDefinitionId:
          type: integer
          format: int64
        chargeName:
          type: string
        customerId:
          type: integer
          format: int64
        overrideAmount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PageableObject:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageCustomerResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/PageableObject'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/CustomerResponse'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        empty:
          type: boolean
    CustomerSearchCriteria:
      type: object
      properties:
        name:
          type: string
        customerNumber:
          type: string
        taxId:
          type: string
        type:
          type: string
          enum:
          - CORPORATE
          - SME
          - INDIVIDUAL
        status:
          type: string
          enum:
          - PENDING
          - ACTIVE
          - SUSPENDED
          - CLOSED
...