# src/main/java/com/banking/masterdata/dto/request/UpdateChannelRequest.java
package com.banking.masterdata.dto.request;

public class UpdateChannelRequest {

    private String name;

    public UpdateChannelRequest() {
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}